+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/v2/Incidents/resolveIncidentErrorType.ts
